+++ src/shogun2-core/src/main/java/de/terrestris/shogun2/service/AbstractCrudService.java
@@ -40,7 +40,9 @@ public abstract class AbstractCrudService<E extends PersistentObject, D extends 
 	 * @param e
 	 * @return
 	 */
-	@PreAuthorize("isAuthenticated()")
+	@PreAuthorize("hasRole(@configHolder.getSuperAdminRoleName())"
+			+ " or (#e.id == null and hasPermission(#e, 'CREATE'))"
+			+ " or (#e.id != null and hasPermission(#e, 'UPDATE'))")
 	public void saveOrUpdate(E e) {
 		dao.saveOrUpdate(e);
 	}